BEPRO 2.8 (#87)
* Dev 1231 make bounty page consider bounty status (#71)

chaging time schedule in get-bounty-moved-to-open

* Dev 778 bounties leaderboard (#72)

* updating models

* adding bounty module

* using database for leaderboard updates

* adding proposal updates

* review fixes

* review fix

* Dev 1233 show token already saved on the database as options (#74)

* updating models

* fix

* handle isAllowedColumn

* environment match

* Bepro 1290 make oracles changed recalculate disputed states (#75)

* create handle is disputed when TVL changes

* fix get address by nework and type

* adding type to updateCuratorProposalParams

* adding handleIsDisputed to oracles-changed-event

* adding where refuedByBountyOwner

* adding await

* rm filter because have a state != closed

* Bepro 1323 change table of tokens to have column is reward so we dont have dupes on database (#76)

* adding isReward to token model

* adding isReward to change-allowed-tokens-events

* adjusting isReward model

* adjusting validateToken function

* adding logic isTransactional

* adding new columns to network_tokens model

* adjusting removed tokens/network_tokens logic

* adding else-if

* Bepro 1319 change network hero to calculate values from the api instead of querying chain (#85)

* creating coinGecko service

* creating header_information model

* adding header_information model to init models

* creating get-prices-header event

* adding .env currency and enable coingecko

* refactoring logic

* adjusting endpoint to new logic

* creating handle-header-information functions

* adding updateBountiesHeader to created events

* adding headerInformationData to changed events

* creating updateNumberOfNetworkHeader function

* adding updateNetwork to registered event

* fix/logger-error (#86)

don't catch the client?.index error so it flows through to the correct catchers

---------

+++ src/actions/get-change-allowed-tokens-events.ts
@@ -44,7 +44,6 @@ export async function action(query?: EventsQuery): Promise<EventsProcessed> {
       const onRegistry = (address: string) => registryTokens[isTransactional ? "transactional" : "reward"].includes(address)
       const notOnRegistry = (token) => !registryTokens[isTransactional ? "transactional" : "reward"].some(a => a === token)
       const onDatabase = (address: string) => tokens.includes(address);
-      const notOnDatabase = (token) => !dbTokens.some((t) => t.address === token && t.isTransactional === isTransactional);
 
       let result: number[]|string[] = [];
 
@@ -52,17 +51,31 @@ export async function action(query?: EventsQuery): Promise<EventsProcessed> {
         result = await Promise.all(
           tokens
             .filter(onRegistry)
-            .filter(notOnDatabase)
             .map(async (tokenAddress) => {
               try {
                 const erc20 = new ERC20(service.web3Connection, tokenAddress)
                 await erc20.loadContract();
-                await db.tokens.create({
-                  name: await erc20.name(),
-                  symbol: await erc20.symbol(),
-                  address: tokenAddress,
-                  isTransactional
+
+                const [token, created] = await db.tokens.findOrCreate({
+                  where: {
+                    address: tokenAddress
+                  },
+                  defaults: {
+                    name: await erc20.name(),
+                    symbol: await erc20.symbol(),
+                    isAllowed: true,
+                    address: tokenAddress,
+                    isTransactional,
+                    isReward: !isTransactional
+                  }
                 });
+
+                if (!created) {
+                  if(isTransactional) token.isTransactional = true
+                  else if(!isTransactional) token.isReward = true
+                  token.isAllowed = true;
+                  await token.save();
+                }
 
                 return tokenAddress;
               } catch (e: any) {
@@ -77,9 +90,25 @@ export async function action(query?: EventsQuery): Promise<EventsProcessed> {
             .filter(onDatabase)
             .map(address => dbTokens.find(t => t.address === address))
             .map(async (token) => {
-              const removed = await db.network_tokens.destroy({where: {tokenId: token.id}});
-              if (!removed)
-                logger.warn(`${name} Failed to remove ${token.id}`);
+              let removed = 0
+              if(isTransactional) {
+                token.isTransactional = false;
+                await db.network_tokens.update({ isTransactional: false }, { where: { tokenId: token.id }})
+              }
+              if(!isTransactional) {
+                token.isReward = false
+                await db.network_tokens.update({ isReward: false }, { where: { tokenId: token.id }})
+              }
+              if(!token.isReward && !token.isTransactional){
+                token.isAllowed = false;
+                removed = await db.network_tokens.destroy({where: {tokenId: token.id}});
+
+                if (!removed)
+                  logger.warn(`${name} Failed to remove ${token.id}`);
+              } 
+              
+              await token.save();
+
               return removed > 0;
             })
         )
